requirements.txt
Flask
numpy
tensorflow-cpu
requests
